address review comment

+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/ClientMessage.java
@@ -203,7 +203,7 @@ public final class ClientMessage implements OutboundFrame {
     /**
      * @return the number of acks will be send for a request
      */
-    public int getNumberOfBackupAcks() {
+    public byte getNumberOfBackupAcks() {
         return getStartFrame().content[RESPONSE_BACKUP_ACKS_FIELD_OFFSET];
     }
 
@@ -213,8 +213,8 @@ public final class ClientMessage implements OutboundFrame {
      * @param numberOfAcks The value to set in the setNumberOfAcks field.
      * @return The ClientMessage with the new dataOffset field value.
      */
-    public ClientMessage setNumberOfBackupAcks(final int numberOfAcks) {
-        getStartFrame().content[RESPONSE_BACKUP_ACKS_FIELD_OFFSET] = (byte) numberOfAcks;
+    public ClientMessage setNumberOfBackupAcks(final byte numberOfAcks) {
+        getStartFrame().content[RESPONSE_BACKUP_ACKS_FIELD_OFFSET] = numberOfAcks;
         return this;
     }
